+++ docs/changelog.md
@@ -6,6 +6,7 @@
 
   - Return more QC metrics from `pp.make_fragment_file`.
   - Also compute library-level TSSe in `metrics.tsse`.
+  - `pp.make_fragment_file` can now work with 10X BAM files.
 
 ## Release 2.6.4 (released May 28, 2024)